refactor: Main.jsからUserContextにユーザー情報取得処理を移動
